SUBS-593 Subscription settings route refactor to match other settings routes
